reflecting yamashita's comments
